fix linting issues and setup typescripts
