+++ package.json
@@ -7,9 +7,11 @@
     "test": "tests"
   },
   "scripts": {
-    "test": "nyc mocha tests",
+    "test": "npm run build && nyc mocha build/tests",
     "start": "node index.js",
-    "lint": "eslint index.js src/"
+    "lint": "eslint index.ts src/**/*.ts --fix",
+    "clean": "rm -rf build",
+    "build": "npm run clean && tsc"
   },
   "keywords": [
     "Xendit"
@@ -24,11 +26,18 @@
     "swagger-ui-express": "^4.1.6"
   },
   "devDependencies": {
+    "@types/async": "2.0.40",
+    "@types/body-parser": "0.0.33",
+    "@types/express": "4.0.34",
+    "@types/mocha": "^9.0.0",
+    "@types/node": "7.10.11",
     "@typescript-eslint/eslint-plugin": "^4.30.0",
     "@typescript-eslint/parser": "^4.30.0",
+    "concurrently": "3.1.0",
     "eslint": "^6.8.0",
     "mocha": "^6.1.4",
     "nyc": "^15.1.0",
-    "supertest": "^4.0.2"
+    "supertest": "^4.0.2",
+    "typescript": "^3.6.5"
   }
 }
